Move Modbus platform schemas into platform files
Move get_hub() to modbus.py and re-export it from __init__.py as
get_hub as get_hub so existing callers (flexit integration) are unaffected.
Platform files import get_hub directly from .modbus to complete the
import-cycle break.

Co-locate each platform's entity schema with its platform file (binary_sensor,
sensor, switch, light, fan, cover, climate) to keep validation and entity
logic in one place. schemas.py then imports these platform schemas to build
CONFIG_SCHEMA instead of defining them inline.

Move BASE_SCHEMA, BASE_COMPONENT_SCHEMA, BASE_STRUCT_SCHEMA and
BASE_SWITCH_SCHEMA to validators.py (which is import-level 1) so platform
files can import base schemas without creating a circular dependency via
__init__.py → schemas.py → platform → __init__.py.
